Replaced [String] dependency url with [Package.Dependency].
The problem with [String] urls was that inside the Package
specification, we were running a map function that was inserting
major version 2 for all the URLS. This is not the best way to deal.

Reason: What if in some near future we add 1 more dependency?
Will that dependency also take 2 as major version? Likely not.

Solution: The current approach adds the ability to add fully
satisfied dependency without the explict knowledge of mapping over.